Create 1074.cpp
A problem that has motivated many studies is to find the minimum number of pieces of the same type, so as to cover the entire board, or the maximum number of pieces of the same type that can be placed without attacking each other, on a clasic cheesboard of 8 × 8 or of another size. The great mathematician Carl F. Gauss, the greatest genius of the modern era, became interested in the "eight queens problem" and found only 72 solutions. All these solutions are obtained for 12 basic locations, rotations and reflections. Your goal is to find all posible solutions for a board of N x N size [1 <= N <= 12].